Added seentracker to globalVar
