+++ lib/src/header_parser/includer.dart
@@ -2,20 +2,13 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:ffigen/src/code_generator.dart';
 import 'data.dart';
 
 /// Utility functions to check whether a binding should be parsed or not
 /// based on filters.
 
-// Stores binding names already scene. Mp key is same as their original name.
-Map<String, Struc> _structs = {};
-Map<String, Func> _functions = {};
-Map<String, EnumClass> _enumClass = {};
-Map<String, String> _macros = {};
-
 bool shouldIncludeStruct(String name) {
-  if (_structs.containsKey(name) || name == '') {
+  if (seenTracker.isSeenStruc(name) || name == '') {
     return false;
   } else if (config.structDecl == null ||
       config.structDecl.shouldInclude(name)) {
@@ -26,7 +19,7 @@ bool shouldIncludeStruct(String name) {
 }
 
 bool shouldIncludeFunc(String name) {
-  if (_functions.containsKey(name) || name == '') {
+  if (seenTracker.isSeenFunc(name) || name == '') {
     return false;
   } else if (config.functionDecl == null ||
       config.functionDecl.shouldInclude(name)) {
@@ -37,7 +30,7 @@ bool shouldIncludeFunc(String name) {
 }
 
 bool shouldIncludeEnumClass(String name) {
-  if (_enumClass.containsKey(name) || name == '') {
+  if (seenTracker.isSeenEnumClass(name) || name == '') {
     return false;
   } else if (config.enumClassDecl == null ||
       config.enumClassDecl.shouldInclude(name)) {
@@ -48,7 +41,7 @@ bool shouldIncludeEnumClass(String name) {
 }
 
 bool shouldIncludeMacro(String name) {
-  if (_macros.containsKey(name) || name == '') {
+  if (seenTracker.isSeenMacro(name) || name == '') {
     return false;
   } else if (config.macroDecl == null || config.macroDecl.shouldInclude(name)) {
     return true;
@@ -75,52 +68,4 @@ bool shouldIncludeRootCursor(String sourceFile) {
   }
 
   return _headerCache[sourceFile];
-}
-
-bool isSeenStruc(String originalName) {
-  return _structs.containsKey(originalName);
-}
-
-void addStrucToSeen(String originalName, Struc struc) {
-  _structs[originalName] = struc;
-}
-
-Struc getSeenStruc(String originalName) {
-  return _structs[originalName];
-}
-
-bool isSeenFunc(String originalName) {
-  return _functions.containsKey(originalName);
-}
-
-void addFuncToSeen(String originalName, Func func) {
-  _functions[originalName] = func;
-}
-
-Func getSeenFunc(String originalName) {
-  return _functions[originalName];
-}
-
-bool isSeenEnumClass(String originalName) {
-  return _enumClass.containsKey(originalName);
-}
-
-void addEnumClassToSeen(String originalName, EnumClass enumClass) {
-  _enumClass[originalName] = enumClass;
-}
-
-EnumClass getSeenEnumClass(String originalName) {
-  return _enumClass[originalName];
-}
-
-bool isSeenMacro(String originalName) {
-  return _macros.containsKey(originalName);
-}
-
-void addMacroToSeen(String originalName, String macro) {
-  _macros[originalName] = macro;
-}
-
-String getSeenMacro(String originalName) {
-  return _macros[originalName];
 }
